moved progress to own file
